Add errors6 solution
I needed a little help with this one, but what finally unlocked it was
realizing that the `unwrap` was not needed once we were mapping the
Result<T, E> returned by `parse`. `unwrap` itself only pulls out the
type from the Result when it is Ok and is apparently discouraged. The
map_err with the ? at the end allowed us to propagate our new combined
error if that was the result (the test that is initially failing that
needs to be fixed), but keep the value of x if there was not an error
(some of the other test cases).