First shot at rustlings (#1)
* Completed through quiz1.rs

* Add move_semantics exercises - took some work to get these at first

* Arrays are immutable! Use them if the size won't change

* Add structs exercises - these made much more sense this time

* Add enums - very cool, though the last one was a little confusing

* Add modules exercises

* Add collections exercises - interesting stuff

* Add strings exercises

* Add quiz2 on strings

This really helped me understand differences between String and &str
much better.

* Add error_handling exercises - tricky and interesting stuff

* Okay, finished errors 5 after looking up help

* Add errors6 solution

I needed a little help with this one, but what finally unlocked it was
realizing that the `unwrap` was not needed once we were mapping the
Result<T, E> returned by `parse`. `unwrap` itself only pulls out the
type from the Result when it is Ok and is apparently discouraged. The
map_err with the ? at the end allowed us to propagate our new combined
error if that was the result (the test that is initially failing that
needs to be fixed), but keep the value of x if there was not an error
(some of the other test cases).

* First two generics exercises

* Okay, figured out generics3 thanks to the compiler

* Well, I got one more